requirements.txt
autogen-agentchat[llm,gemini]~=0.4
autogen-ext[openai]
playwright
streamlit
google-genai
google-generativeai
langgraph
langchain_core
langchain-google-genai
langchain_community
langchain
ipython
langchain_salesforce
langchain-mongodb
pymongo
fastapi
requests
python-dotenv
faiss-cpu
pypdf